# docs/openapi-public.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts": {
      "post": {
        "operationId": "AccountsController_createAccount",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "operationId": "AccountsController_listAccounts",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "industry",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentAccountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/stats": {
      "get": {
        "operationId": "AccountsController_getAccountStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_getAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "operationId": "AccountsController_updateAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      },
      "delete": {
        "operationId": "AccountsController_deleteAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/hierarchy": {
      "get": {
        "operationId": "AccountsController_getAccountHierarchy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/revenue": {
      "get": {
        "operationId": "AccountsController_getAccountRevenue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/bulk/update": {
      "post": {
        "operationId": "AccountsController_bulkUpdate",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/bulk/delete": {
      "post": {
        "operationId": "AccountsController_bulkDelete",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/bulk/assign": {
      "post": {
        "operationId": "AccountsController_bulkAssign",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/meetings/account/{accountId}": {
      "get": {
        "operationId": "MeetingsController_getMeetingsForAccount",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Meetings"
        ]
      }
    },
...
    "/api/activities": {
      "post": {
        "operationId": "ActivitiesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      },
      "get": {
        "operationId": "ActivitiesController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "leadId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opportunityId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/activities/stats": {
      "get": {
        "operationId": "ActivitiesController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/activities/count": {
      "get": {
        "operationId": "ActivitiesController_getActivityCount",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/activities/timeline/{recordType}/{recordId}": {
      "get": {
        "operationId": "ActivitiesController_getTimeline",
        "parameters": [
          {
            "name": "recordType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/activities/{id}": {
      "get": {
        "operationId": "ActivitiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
    "/api/activities/{id}/extract-insights": {
      "post": {
        "operationId": "ActivitiesController_extractInsights",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Activities"
        ]
      }
    },
...
    "/api/orders/account/{accountId}": {
      "get": {
        "operationId": "OrdersController_getByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/territories/{id}/accounts/{accountId}": {
      "delete": {
        "operationId": "TerritoriesController_removeAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Territories"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalesOS API",
    "description": "SalesOS is an AI-powered Sales CRM and Revenue Intelligence platform. This API provides programmatic access to CRM data, sales coaching, CPQ, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://www.salesos.org/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    },
...
  }
}